Solucion de problemas con el SecurityConfig - JWTAuthentication

+++ src/main/java/com/example/aereopuerto/auth/dto/UserDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserDTO {
-    private Long cliente_id;
+    private Integer cliente_id;
     private String email;
     private String telefono;
     private String password;
