+++ src/main/java/com/example/aereopuerto/auth/service/JwtService.java
@@ -39,7 +39,7 @@ public class JwtService {
         return claimsResolver.apply(claims);
     }
 
-    public String generateToken(UserDetails userDetails, Long userId) {
+    public String generateToken(UserDetails userDetails, Integer userId) {
         Map<String, Object> extraClaims = new HashMap<>();
         extraClaims.put("userId", userId);
         return buildToken(extraClaims, userDetails);
